Refactor ID types from int to Long pour tout les model

+++ src/main/java/ma/cabinetplus/dao/MedecinDAO.java
@@ -3,6 +3,6 @@ package ma.cabinetplus.dao;
 import ma.cabinetplus.model.Medecin;
 import java.util.List;
 
-public interface MedecinDAO extends GenericDAO<Medecin, Integer> {
+public interface MedecinDAO extends GenericDAO<Medecin, Long> {
     Medecin trouverParUsername(String username);
 }
